Migrate to custom transpiler
Motivation: The `inline` function supplied by ganja.js uses regex and has bugs with how it handles comments and semi-colons. Using an AST based parser (esprima, via recast) helps avoid these issues. There is also potental to create a babel plugin so that this notation could be supported in code bases with the addition of a compilation state.

Note: Not all cases are considered. Custom basises are not supported, nor are custom Cayley tables. It's more work, I don't fully understand it, and it's beyond the scope of how I intend to use this project.

Future: There is a performance hit in looking up the basis in each operation call. Alternatively, and perferably, it would perhaps make more sense to keep the Algebra consistent across a code base. Hence, the basis would be known at transpile time instead of at runtime as is current. Another consideration could be using dataflow analysis to simplify expressions as much as possible at transpile time, trading transpilation time for better runtime performance. I assume this would be a considerable undertaking for not much gain at this stage though.